+++ demisto_sdk/commands/common/hook_validations/pack_unique_files.py
@@ -182,7 +182,7 @@ class PackUniqueFilesValidator(BaseValidator):
             self._is_price_changed(),
             self._is_approved_tags(),
             self._is_valid_support_type(),
-            self.is_right_usage_of_usecase_tag()
+            self.is_right_usage_of_usecase_tag(),
         ]):
             if self.should_version_raise:
                 return self.validate_version_bump()
@@ -361,14 +361,14 @@ class PackUniqueFilesValidator(BaseValidator):
             pack_meta_file_content = json.loads(self._read_file_content(self.pack_meta_file))
 
             if "Use Case" in pack_meta_file_content['tags']:
-                playbooks_path = self.pack_path + "Playbooks"
-                incidents_path = self.pack_path + "IncidentTypes"
-                layouts_path = self.pack_path + "Layouts"
+                playbooks_path = os.path.join(self.pack_path, "Playbooks")
+                incidents_path = os.path.join(self.pack_path, "IncidentsTypes")
+                layouts_path = os.path.join(self.pack_path, "Layouts")
 
                 answers = [
                     os.path.exists(playbooks_path) and len(os.listdir(playbooks_path)) != 0,
                     os.path.exists(incidents_path) and len(os.listdir(incidents_path)) != 0,
-                    os.path.exists(layouts_path) and len(os.listdir(layouts_path)) != 0
+                    os.path.exists(layouts_path) and len(os.listdir(layouts_path)) != 0,
                 ]
                 if not any(answers):
                     if self._add_error(Errors.is_wrong_usage_of_usecase_tag(), self.pack_meta_file):