+++ demisto_sdk/commands/lint/tests/test_pylint_plugin/xsoar_level_checker_test.py
@@ -164,6 +164,74 @@ class TestTypeAnnotationsChecker(pylint.testutils.CheckerTestCase):
         with self.assertNoMessages():
             self.checker.visit_functiondef(node)
 
+    def test_not_implemented_error_exists_in_if_clause(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - Main function that in the if clause raises a NotImplementedError.
+        Then:
+            - Ensure that the there was not message added to the checker.
+        """
+        node = astroid.extract_node("""
+            def main() -> bool:
+                try:
+                    if command not in commands:
+                        raise NotImplementedError("this command was not implemented")
+                    else:
+                        return True
+                except Exception:
+                    pass
+        """)
+        assert node is not None
+        with self.assertNoMessages():
+            self.checker.visit_functiondef(node)
+
+    def test_not_implemented_error_exists_not_inside_else_if_clauses(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - Main function that raises a NotImplementedError outside of if or else clauses.
+        Then:
+            - Ensure that the there was not message added to the checker.
+        """
+        node = astroid.extract_node("""
+            def main() -> bool:
+                try:
+                    raise NotImplementedError("this command was not implemented")
+                except Exception:
+                    pass
+        """)
+        assert node is not None
+        with self.assertNoMessages():
+            self.checker.visit_functiondef(node)
+
+    def test_not_implemented_error_exists_inside_elif_clause(self):
+        """
+        Given:
+            - String of a code part which is being examined by pylint plugin.
+        When:
+            - Main function that in the elif clause raises a NotImplementedError.
+        Then:
+            - Ensure that the there was not message added to the checker.
+        """
+        node = astroid.extract_node("""
+            def main() -> bool:
+                try:
+                    if True:
+                        return True
+                    elif command not in commands:
+                        raise NotImplementedError("this command was not implemented")
+                    else:
+                        return True
+                except Exception:
+                    pass
+        """)
+        assert node is not None
+        with self.assertNoMessages():
+            self.checker.visit_functiondef(node)
+
     def test_not_implemented_error_doesnt_exists_on_Script(self):
         """
         Given: